Moved neubot/net to neubot/lib_net.
