+++ neubot/bittorrent/__init__.py
@@ -37,7 +37,7 @@ if __name__ == "__main__":
 from neubot.bittorrent.peer import PeerNeubot
 from neubot.bittorrent.server import ServerPeer
 from neubot.http.server import HTTP_SERVER
-from neubot.net.poller import POLLER
+from neubot.lib_net.poller import POLLER
 
 from neubot.bittorrent import config
 from neubot.config import CONFIG
