+++ neubot/bittorrent/stream.py
@@ -26,7 +26,7 @@ import logging
 
 from neubot.bittorrent.config import MAXMESSAGE
 
-from neubot.net.stream import Stream
+from neubot.lib_net.stream import Stream
 
 # Available msgs
 MESSAGES = (CHOKE, UNCHOKE, INTERESTED, NOT_INTERESTED, HAVE, BITFIELD,
