+++ neubot/http/server.py
@@ -30,8 +30,8 @@ from neubot.http.stream import ERROR
 from neubot.http.message import Message
 from neubot.http.stream import nextstate
 from neubot.http.stream import StreamHTTP
-from neubot.net.stream_handler import StreamHandler
-from neubot.net.poller import POLLER
+from neubot.lib_net.stream_handler import StreamHandler
+from neubot.lib_net.poller import POLLER
 
 from neubot import log
 from neubot import utils
