+++ neubot/lib_net/stream.py
