+++ neubot/main.py
@@ -33,7 +33,7 @@ if __name__ == "__main__":
 from neubot.http.message import Message
 from neubot.http.server import HTTP_SERVER
 from neubot.http.server import ServerHTTP
-from neubot.net.poller import POLLER
+from neubot.lib_net.poller import POLLER
 
 from neubot.negotiate.server import NEGOTIATE_SERVER
