+++ neubot/raw_test/raw_srvr.py
@@ -39,7 +39,7 @@ if __name__ == '__main__':
 from .brigade import Brigade
 from .defer import Deferred
 from .handler import Handler
-from neubot.net.poller import POLLER
+from neubot.lib_net.poller import POLLER
 from .raw_defs import AUTH_LEN
 from .raw_defs import EMPTY_MESSAGE
 from .raw_defs import FAKEAUTH
